# contracts/openapi/base.yaml
openapi: 3.0.3
info:
  title: Spider API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/product-docs:
    get:
      summary: List product documents
      parameters:
        - in: query
          name: projectId
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocList"
    post:
      summary: Create product document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDocRequest"
      responses:
        "200":
          description: Product document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocDetail"
  /api/v1/product-docs/{docId}:
    get:
      summary: Get product document detail
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocDetail"
    put:
      summary: Update product document draft
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductDocRequest"
      responses:
        "200":
          description: Product document updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocDetail"
...
  /api/v1/product-docs/{docId}/versions:
    get:
      summary: List product document versions
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product document version list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocVersionList"
  /api/v1/product-docs/{docId}/versions/{versionId}:
    get:
      summary: Get product document version detail
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: versionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product document version detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocVersion"
  /api/v1/product-docs/{docId}/rollback:
    post:
      summary: Roll back to an existing version and publish as a new version
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RollbackProductDocRequest"
      responses:
        "200":
          description: Product document rolled back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocDetail"
  /api/v1/product-docs/{docId}/ai-chat/messages:
    get:
      summary: List product document AI chat messages
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product document AI chat messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocAiMessageList"
    post:
      summary: Send an AI instruction message for product document revision
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDocAiMessageRequest"
      responses:
        "200":
          description: AI revision candidate generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocAiGenerate"
  /api/v1/product-docs/{docId}/ai-chat/messages/stream:
    post:
      summary: Stream AI revision for current product document draft
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDocAiMessageRequest"
      responses:
        "200":
          description: Server-Sent Events stream, emits `delta` and `done` events
          content:
            text/event-stream:
              schema:
                type: string
  /api/v1/product-docs/{docId}/ai-revisions:
    get:
      summary: List product document AI revision candidates
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [pending, confirmed, rejected]
      responses:
        "200":
          description: Product document AI revision list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocRevisionList"
  /api/v1/product-docs/{docId}/ai-revisions/{revisionId}/confirm:
    post:
      summary: Confirm AI revision and publish as a new product document version
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: revisionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmProductDocRevisionRequest"
      responses:
        "200":
          description: AI revision confirmed and version published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocDetail"
  /api/v1/product-docs/{docId}/ai-revisions/{revisionId}/reject:
    post:
      summary: Reject AI revision candidate
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: revisionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectProductDocRevisionRequest"
      responses:
        "200":
          description: AI revision rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductDocRevision"
  /api/v1/chat-sessions:
    get:
      summary: List current actor chat sessions
      parameters:
        - in: query
          name: projectId
          required: false
          schema:
            type: string
            format: uuid
        - in: query
          name: role
          required: false
          schema:
            type: string
            enum: [product, backend, frontend, test]
      responses:
        "200":
          description: Chat session list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseChatSessionList"
    post:
      summary: Create private chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatSessionRequest"
      responses:
        "200":
          description: Chat session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseChatSession"
  /api/v1/chat-sessions/{sessionId}/publish:
    post:
      summary: Create requirement package draft from private chat session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequirementHandoffRequest"
      responses:
        "200":
          description: Requirement handoff published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRequirementHandoff"
  /api/v1/chat-sessions/{sessionId}/messages:
    get:
      summary: List private chat messages for selected session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Chat message list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseChatMessageList"
    post:
      summary: Send a message to chat session and get AI reply
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatMessageRequest"
      responses:
        "200":
          description: AI reply message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseChatMessage"
  /api/v1/chat-sessions/{sessionId}/ai-draft:
    post:
      summary: Generate AI draft handoff package from session context and code impact analysis
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAiDraftRequest"
      responses:
        "200":
          description: AI draft package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAiDraft"
...
components:
  schemas:
...
    ApiResponseProductDocDetail:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/ProductDocDetail"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseProductDocVersion:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/ProductDocVersion"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseChatSession:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/ChatSession"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseRequirementHandoff:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/RequirementHandoff"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseChatMessage:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/ChatMessage"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseAiDraft:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/AiDraft"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    ApiResponseProductDocList:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductDoc"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseProductDocVersionList:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductDocVersion"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseProductDocAiMessageList:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductDocAiMessage"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseProductDocAiGenerate:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/ProductDocAiGenerate"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseProductDocRevision:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/ProductDocRevision"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseProductDocRevisionList:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductDocRevision"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseChatSessionList:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ChatSession"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    ApiResponseChatMessageList:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    ProductDoc:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
        ownerActor:
          type: string
        currentVersionId:
          type: string
          format: uuid
        currentVersionNo:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductDocVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        docId:
          type: string
          format: uuid
        versionNo:
          type: integer
        parentVersionId:
          type: string
          format: uuid
        content:
          type: string
        changeSummary:
          type: string
        sourceType:
          type: string
          enum: [manual, ai_confirm, rollback]
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    ProductDocDetail:
      allOf:
        - $ref: "#/components/schemas/ProductDoc"
        - type: object
          properties:
            draftContent:
              type: string
            currentVersion:
              $ref: "#/components/schemas/ProductDocVersion"
    ProductDocAiMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        docId:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    ProductDocRevision:
      type: object
      properties:
        id:
          type: string
          format: uuid
        docId:
          type: string
          format: uuid
        sourceVersionId:
          type: string
          format: uuid
        baseContent:
          type: string
        candidateContent:
          type: string
        instruction:
          type: string
        changeSummary:
          type: string
        status:
          type: string
          enum: [pending, confirmed, rejected]
        modelProvider:
          type: string
        modelName:
          type: string
        createdBy:
          type: string
        confirmedBy:
          type: string
        confirmedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductDocAiGenerate:
      type: object
      properties:
        assistantMessage:
          $ref: "#/components/schemas/ProductDocAiMessage"
        revision:
          $ref: "#/components/schemas/ProductDocRevision"
    ChatSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        ownerActor:
          type: string
        role:
          type: string
          enum: [product, backend, frontend, test]
        title:
          type: string
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RequirementHandoff:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        sourceSessionId:
          type: string
          format: uuid
        version:
          type: integer
        title:
          type: string
        requirementSummary:
          type: string
        acceptanceCriteria:
          type: string
        impactScope:
          type: string
        priority:
          type: string
          enum: [P0, P1, P2, P3]
        targetRole:
          type: string
          enum: [product, backend, frontend, test]
        status:
          type: string
          enum: [draft, in_review, published, accepted, in_development, in_testing, done, rejected]
        publishedBy:
          type: string
        publishedAt:
          type: string
          format: date-time
        acceptedBy:
          type: string
        acceptedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    AiDraft:
      type: object
      properties:
        title:
          type: string
        requirementSummary:
          type: string
        acceptanceCriteria:
          type: string
        impactScope:
          type: string
        priority:
          type: string
          enum: [P0, P1, P2, P3]
        targetRole:
          type: string
          enum: [product, backend, frontend, test]
        suggestedTasks:
          type: array
          items:
            type: string
        impactedFiles:
          type: array
          items:
            type: string
        riskHints:
          type: array
          items:
            type: string
        testHints:
          type: array
          items:
            type: string
        aiReply:
          type: string
        provider:
          type: string
        model:
          type: string
...
    CreateProductDocRequest:
      type: object
      required: [projectId, title]
      properties:
        projectId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        initialContent:
          type: string
    UpdateProductDocRequest:
      type: object
      required: [title, draftContent, status]
      properties:
        title:
          type: string
          maxLength: 200
        draftContent:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
...
    RollbackProductDocRequest:
      type: object
      required: [targetVersionId]
      properties:
        targetVersionId:
          type: string
          format: uuid
        changeSummary:
          type: string
          maxLength: 500
    CreateProductDocAiMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 4000
    ConfirmProductDocRevisionRequest:
      type: object
      properties:
        changeSummary:
          type: string
          maxLength: 500
    RejectProductDocRevisionRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
...
    CreateChatSessionRequest:
      type: object
      required: [projectId, title]
      properties:
        projectId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
...
    CreateChatMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 8000
    GenerateAiDraftRequest:
      type: object
      properties:
        requirementInput:
          type: string
        priority:
          type: string
          enum: [P0, P1, P2, P3]
          default: P2
        targetRole:
          type: string
          enum: [product, backend, frontend, test]
          default: backend
...
    PublishRequirementHandoffRequest:
      type: object
      required: [title, requirementSummary, acceptanceCriteria, priority, targetRole]
      properties:
        title:
          type: string
          maxLength: 200
        requirementSummary:
          type: string
        acceptanceCriteria:
          type: string
        impactScope:
          type: string
        priority:
          type: string
          enum: [P0, P1, P2, P3]
        targetRole:
          type: string
          enum: [product, backend, frontend, test]